+++ frontend/src/metabase/notifications/modals/QuestionAlertListModal/AlertListItem.tsx
@@ -4,12 +4,12 @@ import { msgid, ngettext, t } from "ttag";
 import _ from "underscore";
 
 import { formatCreatorMessage } from "metabase/account/notifications/components/NotificationCard/utils";
-import { getUser } from "metabase/selectors/user";
-import { Box, FixedSizeIcon, Group, Stack, Text } from "metabase/ui";
 import {
   formatNotificationSchedule,
   getNotificationHandlersGroupedByTypes,
-} from "metabase/utils/notifications";
+} from "metabase/notifications/utils";
+import { getUser } from "metabase/selectors/user";
+import { Box, FixedSizeIcon, Group, Stack, Text } from "metabase/ui";
 import { useSelector } from "metabase/utils/redux";
 import { isNotFalsy } from "metabase/utils/types";
 import type {
