+++ frontend/src/metabase/querying/components/QueryVisualization/VisualizationResult.jsx
@@ -9,8 +9,8 @@ import ButtonsS from "metabase/css/components/buttons.module.css";
 import CS from "metabase/css/core/index.css";
 import { isEmbeddingSdk } from "metabase/embedding-sdk/config";
 import { CreateOrEditQuestionAlertModalWithQuestion } from "metabase/notifications/modals/CreateOrEditQuestionAlertModal/CreateOrEditQuestionAlertModal";
+import { ALERT_TYPE_ROWS, getAlertType } from "metabase/notifications/utils";
 import { zoomInRow } from "metabase/query_builder/actions";
-import { ALERT_TYPE_ROWS, getAlertType } from "metabase/utils/notifications";
 import Visualization from "metabase/visualizations/components/Visualization";
 import * as Lib from "metabase-lib";
 import { datasetContainsNoResults } from "metabase-lib/v1/queries/utils/dataset";
